[jerry] can translate qs_target.nesl
